+++ Editor/PropertyDrawers/DirectoryFilesFieldDrawer.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Jerry Lee. All rights reserved. Licensed under the MIT License. See LICENSE in the
+// project root for license information.
+
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -5,10 +8,10 @@ using UniSharper;
 using UnityEditor;
 using UnityEngine;
 
-namespace ChironEditor
+namespace UniSharperEditor
 {
     /// <summary>
-    /// Property drawer for <see cref="DirectoryFilesFieldDrawer"/>.
+    /// Property drawer for <see cref="UniSharper.DirectoryFilesFieldDrawer"/>.
     /// </summary>
     /// <seealso cref="PropertyDrawer"/>
     [CustomPropertyDrawer(typeof(DirectoryFilesFieldAttribute))]