+++ Runtime/PropertyAttributes/DirectoryFilesFieldAttribute.cs
@@ -11,6 +11,19 @@ namespace UniSharper
     public class DirectoryFilesFieldAttribute : PropertyAttribute
     {
         /// <summary>
+        /// Initializes a new instance of the <see cref="DirectoryFilesFieldAttribute"/> class.
+        /// </summary>
+        /// <param name="directoryPath">The path of target directory. </param>
+        /// <param name="searchPattern">The search pattern. </param>
+        /// <param name="searchOption">The search option. </param>
+        public DirectoryFilesFieldAttribute(string directoryPath, string searchPattern, SearchOption searchOption = SearchOption.TopDirectoryOnly)
+        {
+            DirectoryPath = Path.Combine(Directory.GetCurrentDirectory(), directoryPath);
+            SearchPattern = searchPattern;
+            SearchOption = searchOption;
+        }
+        
+        /// <summary>
         /// The path of target directory.
         /// </summary>
         public string DirectoryPath { get; }
@@ -24,18 +37,5 @@ namespace UniSharper
         /// The search option.
         /// </summary>
         public SearchOption SearchOption { get; }
-        
-        /// <summary>
-        /// Initialize an instance of <see cref="DirectoryFilesFieldAttribute"/>.
-        /// </summary>
-        /// <param name="directoryPath">The path of target directory. </param>
-        /// <param name="searchPattern">The search pattern. </param>
-        /// <param name="searchOption">The search option. </param>
-        public DirectoryFilesFieldAttribute(string directoryPath, string searchPattern, SearchOption searchOption = SearchOption.TopDirectoryOnly)
-        {
-            DirectoryPath = Path.Combine(Directory.GetCurrentDirectory(), directoryPath);
-            SearchPattern = searchPattern;
-            SearchOption = searchOption;
-        }
     }
 }