+++ Runtime/PropertyAttributes/LayerFieldAttribute.cs
@@ -9,12 +9,11 @@ namespace UniSharper
     /// <summary>
     /// Represents the layer field that you select layer name from Tags and Layers settings of Project in the Inspector window.
     /// </summary>
-    /// <seealso cref="PropertyAttribute"/>
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
     public class LayerFieldAttribute : PropertyAttribute
     {
         /// <summary>
-        /// Constructor.
+        /// Initializes a new instance of the <see cref="LayerFieldAttribute"/> class.
         /// </summary>
         /// <param name="useDefaultTagFieldDrawer">Determine whether use the default layer filed drawer.</param>
         public LayerFieldAttribute(bool useDefaultLayerFieldDrawer = true)