+++ Runtime/PropertyAttributes/ReadOnlyFieldAttribute.cs
@@ -9,7 +9,6 @@ namespace UniSharper
     /// <summary>
     /// Represents the readonly field that you can not change the value of field in the Inspector window.
     /// </summary>
-    /// <seealso cref="PropertyAttribute"/>
     [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
     public class ReadOnlyFieldAttribute : PropertyAttribute
     {